wrongly casting to (unsigned int)

+++ tests/strtoll-driver.c
@@ -24,8 +24,8 @@ int main(void)
 	printf("123 ====\n");
 	char *s5 = "123";
 	ans = strtoll(s5, &s5, 10);
-	printf("%lld\n", (unsigned int)ans);
-	printf("%lld\n", 123);
+	printf("%lld\n", ans);
+	printf("%lld\n", 123ll);
 	printf("123 ====\n");
 
 	printf("testing strtoll's ability to handle full numerical ranges..\n");
